refactor(ai-agent): split stream_event.rs into protocol/events/
