+++ crates/aionui-ai-agent/src/protocol/events/mod.rs
@@ -1,0 +1,445 @@
+pub mod permission;
+pub mod session_updates;
+pub mod tool_call;
+pub mod translate;
+
+use serde::{Deserialize, Serialize};
+
+pub use permission::{
+    AcpPermissionEventData, AcpPermissionOptionData, AcpPermissionOptionKind, AcpPermissionRequestData,
+    AcpPermissionToolCall,
+};
+pub use session_updates::{
+    AgentStatusEventData, AvailableCommandsEventData, CronTriggerEventData, PlanEventData, SkillSuggestEventData,
+    ThinkingEventData,
+};
+pub use tool_call::{
+    AcpToolCallContentItem, AcpToolCallEventData, AcpToolCallKind, AcpToolCallLocationItem,
+    AcpToolCallSessionUpdateKind, AcpToolCallStatus, AcpToolCallTextBlock, AcpToolCallTextBlockType,
+    AcpToolCallUpdateData, ToolCallEventData, ToolCallStatus, ToolGroupEntry,
+};
+pub use translate::{permission_request_to_event_data, session_notification_to_chunks, session_notification_to_events};
+
+/// Events emitted by an Agent during a message processing turn.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(tag = "type", content = "data", rename_all = "snake_case")]
+pub enum AgentStreamEvent {
+    Start(StartEventData),
+    #[serde(rename = "content")]
+    Text(TextEventData),
+    Tips(TipsEventData),
+    ToolCall(ToolCallEventData),
+    AcpToolCall(AcpToolCallEventData),
+    ToolGroup(Vec<ToolGroupEntry>),
+    AgentStatus(AgentStatusEventData),
+    Thinking(ThinkingEventData),
+    Plan(PlanEventData),
+    Permission(serde_json::Value),
+    AcpPermission(AcpPermissionEventData),
+    SkillSuggest(SkillSuggestEventData),
+    CronTrigger(CronTriggerEventData),
+    AcpModelInfo(serde_json::Value),
+    AcpModeInfo(serde_json::Value),
+    AcpConfigOption(serde_json::Value),
+    AcpSessionInfo(serde_json::Value),
+    AcpContextUsage(serde_json::Value),
+    SlashCommandsUpdated(serde_json::Value),
+    AvailableCommands(AvailableCommandsEventData),
+    Finish(FinishEventData),
+    Error(ErrorEventData),
+    System(serde_json::Value),
+    RequestTrace(serde_json::Value),
+    SessionAssigned(SessionAssignedEventData),
+}
+
+/// Data for the `Start` event.
+#[derive(Debug, Clone, Default, Serialize, Deserialize)]
+pub struct StartEventData {
+    #[serde(default)]
+    pub session_id: Option<String>,
+}
+
+/// Data for the `SessionAssigned` event.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct SessionAssignedEventData {
+    pub session_id: String,
+}
+
+/// Data for the `Text` event.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct TextEventData {
+    pub content: String,
+}
+
+/// Data for the `Tips` event.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct TipsEventData {
+    pub content: String,
+    #[serde(rename = "type")]
+    pub tip_type: TipType,
+}
+
+/// Severity level for a tip event.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum TipType {
+    Error,
+    Success,
+    Warning,
+}
+
+/// Data for the `Finish` event.
+#[derive(Debug, Clone, Default, Serialize, Deserialize)]
+pub struct FinishEventData {
+    #[serde(default)]
+    pub session_id: Option<String>,
+}
+
+/// Data for the `Error` event.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ErrorEventData {
+    pub message: String,
+    #[serde(default)]
+    pub code: Option<String>,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use agent_client_protocol::schema::{
+        ContentBlock, ContentChunk, PermissionOption, PermissionOptionKind as SdkPermissionOptionKind,
+        RequestPermissionRequest, SessionNotification, SessionUpdate, ToolCall as SdkToolCall,
+        ToolCallStatus as SdkToolCallStatus, ToolCallUpdate as SdkToolCallUpdate, ToolCallUpdateFields,
+        ToolKind as SdkToolKind,
+    };
+    use serde_json::json;
+
+    #[test]
+    fn text_event_roundtrip() {
+        let event = AgentStreamEvent::Text(TextEventData {
+            content: "Hello world".into(),
+        });
+        let json = serde_json::to_value(&event).unwrap();
+        assert_eq!(json["type"], "content");
+        assert_eq!(json["data"]["content"], "Hello world");
+
+        let parsed: AgentStreamEvent = serde_json::from_value(json).unwrap();
+        if let AgentStreamEvent::Text(data) = parsed {
+            assert_eq!(data.content, "Hello world");
+        } else {
+            panic!("Expected Text event");
+        }
+    }
+
+    #[test]
+    fn tips_event_roundtrip() {
+        let event = AgentStreamEvent::Tips(TipsEventData {
+            content: "Something went wrong".into(),
+            tip_type: TipType::Error,
+        });
+        let json = serde_json::to_value(&event).unwrap();
+        assert_eq!(json["type"], "tips");
+        assert_eq!(json["data"]["type"], "error");
+    }
+
+    #[test]
+    fn tool_call_event_roundtrip() {
+        let event = AgentStreamEvent::ToolCall(ToolCallEventData {
+            call_id: "call-1".into(),
+            name: "read_file".into(),
+            args: json!({ "path": "/tmp/a.txt" }),
+            status: ToolCallStatus::Running,
+            input: None,
+            output: None,
+            description: None,
+        });
+        let json = serde_json::to_value(&event).unwrap();
+        assert_eq!(json["type"], "tool_call");
+        assert_eq!(json["data"]["call_id"], "call-1");
+        assert_eq!(json["data"]["status"], "running");
+    }
+
+    #[test]
+    fn tool_call_event_includes_enriched_fields() {
+        let event = AgentStreamEvent::ToolCall(ToolCallEventData {
+            call_id: "call-1".into(),
+            name: "Glob".into(),
+            args: json!({}),
+            status: ToolCallStatus::Completed,
+            input: Some(json!({ "pattern": "**/*.rs" })),
+            output: Some("src/main.rs\nsrc/lib.rs".into()),
+            description: Some("Search for Rust files".into()),
+        });
+        let json = serde_json::to_value(&event).unwrap();
+        assert_eq!(json["type"], "tool_call");
+        assert_eq!(json["data"]["input"]["pattern"], "**/*.rs");
+        assert_eq!(json["data"]["output"], "src/main.rs\nsrc/lib.rs");
+        assert_eq!(json["data"]["description"], "Search for Rust files");
+    }
+
+    #[test]
+    fn tool_call_event_omits_none_fields() {
+        let event = AgentStreamEvent::ToolCall(ToolCallEventData {
+            call_id: "call-1".into(),
+            name: "Glob".into(),
+            args: json!({}),
+            status: ToolCallStatus::Running,
+            input: None,
+            output: None,
+            description: None,
+        });
+        let json = serde_json::to_value(&event).unwrap();
+        assert!(json["data"].get("input").is_none());
+        assert!(json["data"].get("output").is_none());
+        assert!(json["data"].get("description").is_none());
+    }
+
+    #[test]
+    fn finish_event_roundtrip() {
+        let event = AgentStreamEvent::Finish(FinishEventData {
+            session_id: Some("sess-abc".into()),
+        });
+        let json = serde_json::to_value(&event).unwrap();
+        assert_eq!(json["type"], "finish");
+        assert_eq!(json["data"]["session_id"], "sess-abc");
+    }
+
+    #[test]
+    fn error_event_roundtrip() {
+        let event = AgentStreamEvent::Error(ErrorEventData {
+            message: "timeout".into(),
+            code: Some("E001".into()),
+        });
+        let json = serde_json::to_value(&event).unwrap();
+        assert_eq!(json["type"], "error");
+        assert_eq!(json["data"]["message"], "timeout");
+    }
+
+    #[test]
+    fn start_event_default_session_id() {
+        let event = AgentStreamEvent::Start(StartEventData::default());
+        let json = serde_json::to_value(&event).unwrap();
+        assert_eq!(json["type"], "start");
+        assert_eq!(json["data"]["session_id"], serde_json::Value::Null);
+    }
+
+    #[test]
+    fn tool_group_event_roundtrip() {
+        let entries = vec![
+            ToolGroupEntry {
+                call_id: "c1".into(),
+                name: "read".into(),
+                status: ToolCallStatus::Completed,
+                description: Some("Read file".into()),
+            },
+            ToolGroupEntry {
+                call_id: "c2".into(),
+                name: "write".into(),
+                status: ToolCallStatus::Running,
+                description: None,
+            },
+        ];
+        let event = AgentStreamEvent::ToolGroup(entries);
+        let json = serde_json::to_value(&event).unwrap();
+        assert_eq!(json["type"], "tool_group");
+        let data = json["data"].as_array().unwrap();
+        assert_eq!(data.len(), 2);
+        assert_eq!(data[0]["call_id"], "c1");
+    }
+
+    #[test]
+    fn agent_status_event_roundtrip() {
+        let event = AgentStreamEvent::AgentStatus(AgentStatusEventData {
+            backend: "claude".into(),
+            status: "running".into(),
+            agent_name: Some("default".into()),
+            session_id: None,
+        });
+        let json = serde_json::to_value(&event).unwrap();
+        assert_eq!(json["type"], "agent_status");
+        assert_eq!(json["data"]["backend"], "claude");
+    }
+
+    #[test]
+    fn session_tool_call_maps_to_acp_tool_call_event() {
+        let notif = SessionNotification::new(
+            "sess-1",
+            SessionUpdate::ToolCall(
+                SdkToolCall::new("tool-1", "Terminal")
+                    .kind(SdkToolKind::Execute)
+                    .status(SdkToolCallStatus::Pending)
+                    .raw_input(json!({ "command": "echo hi" })),
+            ),
+        );
+
+        let events = session_notification_to_events(&notif);
+        assert_eq!(events.len(), 1);
+        let json = serde_json::to_value(&events[0]).unwrap();
+        assert_eq!(json["type"], "acp_tool_call");
+        assert_eq!(json["data"]["session_id"], "sess-1");
+        assert_eq!(json["data"]["update"]["sessionUpdate"], "tool_call");
+        assert_eq!(json["data"]["update"]["tool_call_id"], "tool-1");
+        assert_eq!(json["data"]["update"]["title"], "Terminal");
+        assert_eq!(json["data"]["update"]["kind"], "execute");
+        assert_eq!(json["data"]["update"]["rawInput"]["command"], "echo hi");
+    }
+
+    #[test]
+    fn session_tool_call_update_omits_missing_fields_for_frontend_merge() {
+        let notif = SessionNotification::new(
+            "sess-1",
+            SessionUpdate::ToolCallUpdate(SdkToolCallUpdate::new(
+                "tool-1",
+                ToolCallUpdateFields::new().status(SdkToolCallStatus::Completed),
+            )),
+        );
+
+        let events = session_notification_to_events(&notif);
+        assert_eq!(events.len(), 1);
+        let json = serde_json::to_value(&events[0]).unwrap();
+        assert_eq!(json["type"], "acp_tool_call");
+        assert_eq!(json["data"]["update"]["sessionUpdate"], "tool_call_update");
+        assert_eq!(json["data"]["update"]["tool_call_id"], "tool-1");
+        assert_eq!(json["data"]["update"]["status"], "completed");
+        assert!(json["data"]["update"].get("title").is_none());
+        assert!(json["data"]["update"].get("rawInput").is_none());
+    }
+
+    #[test]
+    fn permission_request_maps_to_snake_case_event_data() {
+        let request = RequestPermissionRequest::new(
+            "sess-1",
+            SdkToolCallUpdate::new(
+                "tool-1",
+                ToolCallUpdateFields::new()
+                    .title("Write file")
+                    .kind(SdkToolKind::Edit)
+                    .raw_input(json!({ "file_path": "/tmp/a.txt" })),
+            ),
+            vec![
+                PermissionOption::new("allow", "Allow", SdkPermissionOptionKind::AllowOnce),
+                PermissionOption::new("reject", "Reject", SdkPermissionOptionKind::RejectOnce),
+            ],
+        );
+
+        let event = AgentStreamEvent::AcpPermission(permission_request_to_event_data(&request));
+        let json = serde_json::to_value(&event).unwrap();
+
+        assert_eq!(json["type"], "acp_permission");
+        assert_eq!(json["data"]["session_id"], "sess-1");
+        assert_eq!(json["data"]["tool_call"]["tool_call_id"], "tool-1");
+        assert_eq!(json["data"]["tool_call"]["raw_input"]["file_path"], "/tmp/a.txt");
+        assert_eq!(json["data"]["options"][0]["option_id"], "allow");
+        assert_eq!(json["data"]["options"][0]["kind"], "allow_once");
+        assert!(json["data"].get("toolCall").is_none());
+        assert!(json["data"]["options"][0].get("optionId").is_none());
+    }
+
+    #[test]
+    fn thinking_event_roundtrip() {
+        let event = AgentStreamEvent::Thinking(ThinkingEventData {
+            content: "Analyzing...".into(),
+            subject: Some("code review".into()),
+            duration: Some(1500),
+            status: Some("in_progress".into()),
+        });
+        let json = serde_json::to_value(&event).unwrap();
+        assert_eq!(json["type"], "thinking");
+        assert_eq!(json["data"]["duration"], 1500);
+    }
+
+    #[test]
+    fn agent_message_chunk_maps_to_text_chunk() {
+        let notif = SessionNotification::new(
+            "sess-1",
+            SessionUpdate::AgentMessageChunk(ContentChunk::new(ContentBlock::from("hello world"))),
+        );
+        let chunks = session_notification_to_chunks(&notif);
+        assert_eq!(chunks.len(), 1);
+        match &chunks[0] {
+            AgentStreamChunk::Text { text } => assert_eq!(text, "hello world"),
+            other => panic!("expected Text, got {other:?}"),
+        }
+    }
+
+    #[test]
+    fn agent_thought_chunk_maps_to_thought_chunk() {
+        let notif = SessionNotification::new(
+            "sess-1",
+            SessionUpdate::AgentThoughtChunk(ContentChunk::new(ContentBlock::from("thinking..."))),
+        );
+        let chunks = session_notification_to_chunks(&notif);
+        assert_eq!(chunks.len(), 1);
+        match &chunks[0] {
+            AgentStreamChunk::Thought { content } => assert_eq!(content, "thinking..."),
+            other => panic!("expected Thought, got {other:?}"),
+        }
+    }
+
+    #[test]
+    fn tool_call_maps_to_tool_use_chunk() {
+        let notif = SessionNotification::new(
+            "sess-1",
+            SessionUpdate::ToolCall(
+                SdkToolCall::new("tool-1", "read_file")
+                    .kind(SdkToolKind::Read)
+                    .status(SdkToolCallStatus::Pending)
+                    .raw_input(json!({ "path": "/tmp/a.txt" })),
+            ),
+        );
+        let chunks = session_notification_to_chunks(&notif);
+        assert_eq!(chunks.len(), 1);
+        match &chunks[0] {
+            AgentStreamChunk::ToolUse { tool_name, input } => {
+                assert_eq!(tool_name, "read_file");
+                assert_eq!(input["path"], "/tmp/a.txt");
+            }
+            other => panic!("expected ToolUse, got {other:?}"),
+        }
+    }
+
+    #[test]
+    fn tool_call_update_without_raw_input_is_skipped() {
+        let notif = SessionNotification::new(
+            "sess-1",
+            SessionUpdate::ToolCallUpdate(SdkToolCallUpdate::new(
+                "tool-1",
+                ToolCallUpdateFields::new().status(SdkToolCallStatus::Completed),
+            )),
+        );
+        let chunks = session_notification_to_chunks(&notif);
+        assert!(chunks.is_empty());
+    }
+
+    #[test]
+    fn tool_call_update_with_raw_input_emits_tool_use() {
+        let notif = SessionNotification::new(
+            "sess-1",
+            SessionUpdate::ToolCallUpdate(SdkToolCallUpdate::new(
+                "tool-1",
+                ToolCallUpdateFields::new()
+                    .title("search")
+                    .raw_input(json!({ "q": "foo" })),
+            )),
+        );
+        let chunks = session_notification_to_chunks(&notif);
+        assert_eq!(chunks.len(), 1);
+        match &chunks[0] {
+            AgentStreamChunk::ToolUse { tool_name, input } => {
+                assert_eq!(tool_name, "search");
+                assert_eq!(input["q"], "foo");
+            }
+            other => panic!("expected ToolUse, got {other:?}"),
+        }
+    }
+
+    #[test]
+    fn plan_update_does_not_produce_chunks() {
+        let notif = SessionNotification::new(
+            "sess-1",
+            SessionUpdate::CurrentModeUpdate(agent_client_protocol::schema::CurrentModeUpdate::new("normal")),
+        );
+        let chunks = session_notification_to_chunks(&notif);
+        assert!(chunks.is_empty());
+    }
+}
