+++ crates/aionui-ai-agent/src/protocol/mod.rs
@@ -1,0 +1,1 @@
+pub mod events;
