+++ crates/aionui-ai-agent/src/stream_event.rs
@@ -1,1063 +1,1 @@
-use agent_client_protocol::schema::{
-    AvailableCommand, ContentBlock, Meta as SdkMeta, PermissionOption, PermissionOptionKind as SdkPermissionOptionKind,
-    RequestPermissionRequest, SessionNotification, SessionUpdate, ToolCallContent as SdkToolCallContent,
-    ToolCallLocation as SdkToolCallLocation, ToolCallStatus as SdkToolCallStatus, ToolCallUpdate as SdkToolCallUpdate,
-    ToolKind as SdkToolKind,
-};
-use aionui_common::{Confirmation, ConfirmationOption};
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
-use tracing::debug;
-
-use crate::types::AgentStreamChunk;
-
-/// Events emitted by an Agent during a message processing turn.
-///
-/// These are parsed from Agent stdout (line-delimited JSON) and forwarded
-/// to the WebSocket layer as `message.stream` events.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(tag = "type", content = "data", rename_all = "snake_case")]
-pub enum AgentStreamEvent {
-    /// Start of a new response turn.
-    Start(StartEventData),
-    /// Incremental text content.
-    #[serde(rename = "content")]
-    Text(TextEventData),
-    /// Tip / notification (error, success, warning).
-    Tips(TipsEventData),
-
-    /// Single tool call status update.
-    ToolCall(ToolCallEventData),
-    /// ACP tool call progress.
-    AcpToolCall(AcpToolCallEventData),
-
-    /// Group of tool calls.
-    ToolGroup(Vec<ToolGroupEntry>),
-    /// Agent status change (backend, status, session info).
-    AgentStatus(AgentStatusEventData),
-    /// Thinking / reasoning trace.
-    Thinking(ThinkingEventData),
-    /// Execution plan.
-    Plan(PlanEventData),
-    /// Generic permission request (non-ACP backends).
-    Permission(serde_json::Value),
-    /// ACP permission request (tool approval).
-    AcpPermission(AcpPermissionEventData),
-    /// Skill suggestion from cron job.
-    SkillSuggest(SkillSuggestEventData),
-    /// Cron trigger notification.
-    CronTrigger(CronTriggerEventData),
-
-    /// ACP model info update.
-    AcpModelInfo(serde_json::Value),
-    /// ACP current session mode update.
-    AcpModeInfo(serde_json::Value),
-    /// ACP session config option update.
-    AcpConfigOption(serde_json::Value),
-    /// ACP session info update (title / timestamps / metadata).
-    AcpSessionInfo(serde_json::Value),
-    /// ACP context usage info.
-    AcpContextUsage(serde_json::Value),
-
-    /// Slash commands updated notification.
-    SlashCommandsUpdated(serde_json::Value),
-    /// Available slash commands update.
-    AvailableCommands(AvailableCommandsEventData),
-
-    /// Response finished.
-    Finish(FinishEventData),
-    /// Error during processing.
-    Error(ErrorEventData),
-    /// System-level message from ACP.
-    System(serde_json::Value),
-    /// Raw request trace (ACP debug info).
-    RequestTrace(serde_json::Value),
-
-    /// Emitted exactly once per conversation, right after the ACP CLI
-    /// returns a fresh `session_id` from `session/new`. Subscribers
-    /// (e.g. `AcpAgentService`) persist the id into `acp_session` so a
-    /// later resume can call `session/load`.
-    ///
-    /// Not emitted on `session/load` — the id in that path is
-    /// already stored.
-    SessionAssigned(SessionAssignedEventData),
-}
-
-/// Data for the `Start` event.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct StartEventData {
-    /// Session ID for this turn (if available).
-    #[serde(default)]
-    pub session_id: Option<String>,
-}
-
-/// Data for the `SessionAssigned` event.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SessionAssignedEventData {
-    /// The CLI-assigned session id, freshly minted by `session/new`.
-    pub session_id: String,
-}
-
-/// Data for the `Text` event.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TextEventData {
-    /// Incremental text content.
-    pub content: String,
-}
-
-/// Data for the `Tips` event.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TipsEventData {
-    /// Tip message content.
-    pub content: String,
-    /// Severity level.
-    #[serde(rename = "type")]
-    pub tip_type: TipType,
-}
-
-/// Severity level for a tip event.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum TipType {
-    Error,
-    Success,
-    Warning,
-}
-
-/// Data for the `ToolCall` event.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ToolCallEventData {
-    pub call_id: String,
-    pub name: String,
-    #[serde(default)]
-    pub args: serde_json::Value,
-    pub status: ToolCallStatus,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub input: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub output: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AcpToolCallEventData {
-    pub session_id: String,
-    pub update: AcpToolCallUpdateData,
-    #[serde(rename = "_meta", skip_serializing_if = "Option::is_none")]
-    pub meta: Option<SdkMeta>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AcpToolCallUpdateData {
-    #[serde(rename = "sessionUpdate")]
-    pub session_update: AcpToolCallSessionUpdateKind,
-    pub tool_call_id: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub status: Option<AcpToolCallStatus>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub title: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub kind: Option<AcpToolCallKind>,
-    #[serde(rename = "rawInput", skip_serializing_if = "Option::is_none")]
-    pub raw_input: Option<Value>,
-    #[serde(rename = "rawOutput", skip_serializing_if = "Option::is_none")]
-    pub raw_output: Option<Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub content: Option<Vec<AcpToolCallContentItem>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub locations: Option<Vec<AcpToolCallLocationItem>>,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum AcpToolCallSessionUpdateKind {
-    ToolCall,
-    ToolCallUpdate,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum AcpToolCallStatus {
-    Pending,
-    InProgress,
-    Completed,
-    Failed,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum AcpToolCallKind {
-    Read,
-    Edit,
-    Execute,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(tag = "type", rename_all = "snake_case")]
-pub enum AcpToolCallContentItem {
-    Content {
-        content: AcpToolCallTextBlock,
-    },
-    Diff {
-        path: String,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        old_text: Option<String>,
-        new_text: String,
-    },
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AcpToolCallTextBlock {
-    #[serde(rename = "type")]
-    pub block_type: AcpToolCallTextBlockType,
-    pub text: String,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum AcpToolCallTextBlockType {
-    Text,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AcpToolCallLocationItem {
-    pub path: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum AcpPermissionEventData {
-    Request(AcpPermissionRequestData),
-    Confirmation(Confirmation),
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AcpPermissionRequestData {
-    #[serde(default)]
-    pub session_id: String,
-    pub tool_call: AcpPermissionToolCall,
-    pub options: Vec<AcpPermissionOptionData>,
-    #[serde(rename = "_meta", skip_serializing_if = "Option::is_none")]
-    pub meta: Option<SdkMeta>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AcpPermissionToolCall {
-    pub tool_call_id: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub status: Option<AcpToolCallStatus>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub title: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub kind: Option<AcpToolCallKind>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub raw_input: Option<Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub raw_output: Option<Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub content: Option<Vec<AcpToolCallContentItem>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub locations: Option<Vec<AcpToolCallLocationItem>>,
-    #[serde(rename = "_meta", skip_serializing_if = "Option::is_none")]
-    pub meta: Option<SdkMeta>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AcpPermissionOptionData {
-    pub option_id: String,
-    pub name: String,
-    pub kind: AcpPermissionOptionKind,
-    #[serde(rename = "_meta", skip_serializing_if = "Option::is_none")]
-    pub meta: Option<SdkMeta>,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum AcpPermissionOptionKind {
-    AllowOnce,
-    AllowAlways,
-    RejectOnce,
-    RejectAlways,
-}
-
-impl AcpPermissionEventData {
-    pub fn as_confirmation(&self) -> Option<Confirmation> {
-        match self {
-            Self::Confirmation(conf) => Some(conf.clone()),
-            Self::Request(req) => Some(req.to_confirmation()),
-        }
-    }
-}
-
-impl AcpPermissionRequestData {
-    pub fn to_confirmation(&self) -> Confirmation {
-        Confirmation {
-            id: self.tool_call.tool_call_id.clone(),
-            call_id: self.tool_call.tool_call_id.clone(),
-            title: self.tool_call.title.clone(),
-            action: None,
-            description: self
-                .tool_call
-                .raw_input
-                .as_ref()
-                .and_then(|raw| raw.get("description").and_then(Value::as_str))
-                .map(ToOwned::to_owned)
-                .unwrap_or_else(|| {
-                    self.tool_call
-                        .raw_input
-                        .as_ref()
-                        .map(Value::to_string)
-                        .unwrap_or_default()
-                }),
-            command_type: self.tool_call.kind.map(|kind| match kind {
-                AcpToolCallKind::Read => "read".to_owned(),
-                AcpToolCallKind::Edit => "edit".to_owned(),
-                AcpToolCallKind::Execute => "execute".to_owned(),
-            }),
-            options: self
-                .options
-                .iter()
-                .map(|opt| ConfirmationOption {
-                    label: opt.name.clone(),
-                    value: Value::String(opt.option_id.clone()),
-                    params: None,
-                })
-                .collect(),
-        }
-    }
-}
-
-/// Status of a tool call.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ToolCallStatus {
-    Running,
-    Completed,
-    Error,
-}
-
-/// A single entry in a `ToolGroup` event.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ToolGroupEntry {
-    pub call_id: String,
-    pub name: String,
-    pub status: ToolCallStatus,
-    #[serde(default)]
-    pub description: Option<String>,
-}
-
-/// Data for the `AgentStatus` event.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AgentStatusEventData {
-    pub backend: String,
-    pub status: String,
-    #[serde(default)]
-    pub agent_name: Option<String>,
-    #[serde(default)]
-    pub session_id: Option<String>,
-}
-
-/// Data for the `Thinking` event.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ThinkingEventData {
-    pub content: String,
-    #[serde(default)]
-    pub subject: Option<String>,
-    #[serde(default)]
-    pub duration: Option<u64>,
-    #[serde(default)]
-    pub status: Option<String>,
-}
-
-/// Data for the `Plan` event.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PlanEventData {
-    #[serde(default)]
-    pub session_id: Option<String>,
-    #[serde(default)]
-    pub entries: Vec<serde_json::Value>,
-}
-
-/// Data for the `AvailableCommands` event.
-///
-/// `commands` carries ACP SDK structs that serialise as camelCase on
-/// their own. Producers keep them typed; the case normalisation to
-/// snake_case happens at the stream-relay layer when the event is
-/// serialised for WebSocket delivery.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AvailableCommandsEventData {
-    pub commands: Vec<AvailableCommand>,
-}
-
-/// Data for the `SkillSuggest` event.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SkillSuggestEventData {
-    #[serde(default)]
-    pub cron_job_id: Option<String>,
-    pub name: String,
-    pub description: String,
-    #[serde(default)]
-    pub skill_content: Option<String>,
-}
-
-/// Data for the `CronTrigger` event.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CronTriggerEventData {
-    pub cron_job_id: String,
-    pub cron_job_name: String,
-    pub triggered_at: i64,
-}
-
-/// Data for the `Finish` event.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct FinishEventData {
-    #[serde(default)]
-    pub session_id: Option<String>,
-}
-
-/// Data for the `Error` event.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ErrorEventData {
-    pub message: String,
-    #[serde(default)]
-    pub code: Option<String>,
-}
-
-// ── SDK SessionNotification → AgentStreamEvent conversion ────────────────────
-
-/// Convert an SDK [`SessionNotification`] into zero or more [`AgentStreamEvent`]s.
-///
-/// Each `SessionUpdate` variant is mapped to the closest existing event type.
-/// Unknown or unmappable variants produce a debug log and are skipped (not
-/// silently swallowed, not panicked).
-pub fn session_notification_to_events(notif: &SessionNotification) -> Vec<AgentStreamEvent> {
-    let session_id = notif.session_id.to_string();
-    let mut events = Vec::new();
-
-    match &notif.update {
-        SessionUpdate::AgentMessageChunk(chunk) => {
-            if let ContentBlock::Text(text) = &chunk.content {
-                events.push(AgentStreamEvent::Text(TextEventData {
-                    content: text.text.clone(),
-                }));
-            }
-        }
-
-        SessionUpdate::AgentThoughtChunk(chunk) => {
-            if let ContentBlock::Text(text) = &chunk.content {
-                events.push(AgentStreamEvent::Thinking(ThinkingEventData {
-                    content: text.text.clone(),
-                    subject: None,
-                    duration: None,
-                    status: Some("in_progress".into()),
-                }));
-            }
-        }
-
-        SessionUpdate::UserMessageChunk(_chunk) => {
-            // User message echoes are not forwarded to the event stream.
-            // The frontend already has the user's message.
-        }
-
-        SessionUpdate::ToolCall(tc) => {
-            events.push(AgentStreamEvent::AcpToolCall(AcpToolCallEventData {
-                session_id,
-                update: AcpToolCallUpdateData {
-                    session_update: AcpToolCallSessionUpdateKind::ToolCall,
-                    tool_call_id: tc.tool_call_id.to_string(),
-                    status: Some(map_sdk_tool_status(&tc.status)),
-                    title: Some(tc.title.clone()),
-                    kind: Some(map_sdk_tool_kind(&tc.kind)),
-                    raw_input: tc.raw_input.clone(),
-                    raw_output: None,
-                    content: map_tool_call_content(&tc.content),
-                    locations: map_tool_call_locations(&tc.locations),
-                },
-                meta: tc.meta.clone(),
-            }));
-        }
-
-        SessionUpdate::ToolCallUpdate(tcu) => {
-            events.push(AgentStreamEvent::AcpToolCall(AcpToolCallEventData {
-                session_id,
-                update: AcpToolCallUpdateData {
-                    session_update: AcpToolCallSessionUpdateKind::ToolCallUpdate,
-                    tool_call_id: tcu.tool_call_id.to_string(),
-                    status: tcu.fields.status.as_ref().map(map_sdk_tool_status),
-                    title: tcu.fields.title.clone(),
-                    kind: tcu.fields.kind.as_ref().map(map_sdk_tool_kind),
-                    raw_input: tcu.fields.raw_input.clone(),
-                    raw_output: tcu.fields.raw_output.clone(),
-                    content: tcu
-                        .fields
-                        .content
-                        .as_ref()
-                        .and_then(|content| map_tool_call_content(content)),
-                    locations: tcu
-                        .fields
-                        .locations
-                        .as_ref()
-                        .and_then(|locations| map_tool_call_locations(locations)),
-                },
-                meta: tcu.meta.clone(),
-            }));
-        }
-
-        SessionUpdate::Plan(plan) => {
-            let entries: Vec<serde_json::Value> = plan
-                .entries
-                .iter()
-                .map(|e| serde_json::to_value(e).unwrap_or_default())
-                .collect();
-
-            events.push(AgentStreamEvent::Plan(PlanEventData {
-                session_id: Some(session_id),
-                entries,
-            }));
-        }
-
-        SessionUpdate::AvailableCommandsUpdate(update) => {
-            events.push(AgentStreamEvent::AvailableCommands(AvailableCommandsEventData {
-                commands: update.available_commands.clone(),
-            }));
-        }
-
-        SessionUpdate::CurrentModeUpdate(update) => {
-            events.push(AgentStreamEvent::AcpModeInfo(
-                serde_json::to_value(update).unwrap_or_default(),
-            ));
-        }
-
-        SessionUpdate::ConfigOptionUpdate(update) => {
-            events.push(AgentStreamEvent::AcpConfigOption(
-                serde_json::to_value(update).unwrap_or_default(),
-            ));
-        }
-
-        SessionUpdate::SessionInfoUpdate(update) => {
-            events.push(AgentStreamEvent::AcpSessionInfo(
-                serde_json::to_value(update).unwrap_or_default(),
-            ));
-        }
-
-        SessionUpdate::UsageUpdate(update) => {
-            events.push(AgentStreamEvent::AcpContextUsage(
-                serde_json::to_value(update).unwrap_or_default(),
-            ));
-        }
-        // Future SDK variants or feature-gated variants — log and skip.
-        _ => {
-            debug!("Unknown SessionUpdate variant received, skipping");
-        }
-    }
-
-    events
-}
-
-/// Convert an SDK [`SessionNotification`] into zero or more [`AgentStreamChunk`]s.
-///
-/// Parallel projection of [`session_notification_to_events`] onto the smaller
-/// [`AgentStreamChunk`] enum consumed by `subscribe_stream()`. Only message/
-/// thought/tool-call updates are projected — handshake-like updates
-/// (mode/model/config/session-info/usage/plan) are skipped since they are not
-/// useful for the wake-timeout watchdog or crash detector that subscribe to
-/// this stream.
-pub fn session_notification_to_chunks(notif: &SessionNotification) -> Vec<AgentStreamChunk> {
-    let mut chunks = Vec::new();
-
-    match &notif.update {
-        SessionUpdate::AgentMessageChunk(chunk) => {
-            if let ContentBlock::Text(text) = &chunk.content {
-                chunks.push(AgentStreamChunk::Text {
-                    text: text.text.clone(),
-                });
-            }
-        }
-        SessionUpdate::AgentThoughtChunk(chunk) => {
-            if let ContentBlock::Text(text) = &chunk.content {
-                chunks.push(AgentStreamChunk::Thought {
-                    content: text.text.clone(),
-                });
-            }
-        }
-        SessionUpdate::ToolCall(tc) => {
-            chunks.push(AgentStreamChunk::ToolUse {
-                tool_name: tc.title.clone(),
-                input: tc.raw_input.clone().unwrap_or(Value::Null),
-            });
-        }
-        SessionUpdate::ToolCallUpdate(tcu) => {
-            // Only emit ToolUse on updates that carry a fresh raw_input —
-            // status-only updates (e.g. Completed with no new input) would
-            // duplicate earlier emits without adding signal for the
-            // watchdog.
-            if let Some(raw_input) = tcu.fields.raw_input.clone() {
-                let tool_name = tcu.fields.title.clone().unwrap_or_default();
-                chunks.push(AgentStreamChunk::ToolUse {
-                    tool_name,
-                    input: raw_input,
-                });
-            }
-        }
-        _ => {}
-    }
-
-    chunks
-}
-
-fn map_sdk_tool_status(sdk: &SdkToolCallStatus) -> AcpToolCallStatus {
-    match sdk {
-        SdkToolCallStatus::Pending => AcpToolCallStatus::Pending,
-        SdkToolCallStatus::InProgress => AcpToolCallStatus::InProgress,
-        SdkToolCallStatus::Completed => AcpToolCallStatus::Completed,
-        SdkToolCallStatus::Failed => AcpToolCallStatus::Failed,
-        _ => AcpToolCallStatus::Pending,
-    }
-}
-
-fn map_sdk_tool_kind(kind: &SdkToolKind) -> AcpToolCallKind {
-    match kind {
-        SdkToolKind::Read | SdkToolKind::Search => AcpToolCallKind::Read,
-        SdkToolKind::Edit | SdkToolKind::Delete | SdkToolKind::Move => AcpToolCallKind::Edit,
-        SdkToolKind::Execute
-        | SdkToolKind::Think
-        | SdkToolKind::Fetch
-        | SdkToolKind::SwitchMode
-        | SdkToolKind::Other
-        | _ => AcpToolCallKind::Execute,
-    }
-}
-
-fn map_sdk_permission_option_kind(kind: SdkPermissionOptionKind) -> AcpPermissionOptionKind {
-    match kind {
-        SdkPermissionOptionKind::AllowOnce => AcpPermissionOptionKind::AllowOnce,
-        SdkPermissionOptionKind::AllowAlways => AcpPermissionOptionKind::AllowAlways,
-        SdkPermissionOptionKind::RejectOnce => AcpPermissionOptionKind::RejectOnce,
-        SdkPermissionOptionKind::RejectAlways => AcpPermissionOptionKind::RejectAlways,
-        _ => AcpPermissionOptionKind::RejectOnce,
-    }
-}
-
-pub fn permission_request_to_event_data(request: &RequestPermissionRequest) -> AcpPermissionEventData {
-    AcpPermissionEventData::Request(AcpPermissionRequestData {
-        session_id: request.session_id.to_string(),
-        tool_call: map_permission_tool_call(&request.tool_call),
-        options: request.options.iter().map(map_permission_option).collect(),
-        meta: request.meta.clone(),
-    })
-}
-
-fn map_permission_tool_call(tool_call: &SdkToolCallUpdate) -> AcpPermissionToolCall {
-    AcpPermissionToolCall {
-        tool_call_id: tool_call.tool_call_id.to_string(),
-        status: tool_call.fields.status.as_ref().map(map_sdk_tool_status),
-        title: tool_call.fields.title.clone(),
-        kind: tool_call.fields.kind.as_ref().map(map_sdk_tool_kind),
-        raw_input: tool_call.fields.raw_input.clone(),
-        raw_output: tool_call.fields.raw_output.clone(),
-        content: tool_call
-            .fields
-            .content
-            .as_ref()
-            .and_then(|content| map_tool_call_content(content)),
-        locations: tool_call
-            .fields
-            .locations
-            .as_ref()
-            .and_then(|locations| map_tool_call_locations(locations)),
-        meta: tool_call.meta.clone(),
-    }
-}
-
-fn map_permission_option(option: &PermissionOption) -> AcpPermissionOptionData {
-    AcpPermissionOptionData {
-        option_id: option.option_id.to_string(),
-        name: option.name.clone(),
-        kind: map_sdk_permission_option_kind(option.kind),
-        meta: option.meta.clone(),
-    }
-}
-
-fn map_tool_call_content(content: &[SdkToolCallContent]) -> Option<Vec<AcpToolCallContentItem>> {
-    let items: Vec<AcpToolCallContentItem> = content
-        .iter()
-        .filter_map(|item| match item {
-            SdkToolCallContent::Content(content) => match &content.content {
-                ContentBlock::Text(text) => Some(AcpToolCallContentItem::Content {
-                    content: AcpToolCallTextBlock {
-                        block_type: AcpToolCallTextBlockType::Text,
-                        text: text.text.clone(),
-                    },
-                }),
-                _ => None,
-            },
-            SdkToolCallContent::Diff(diff) => Some(AcpToolCallContentItem::Diff {
-                path: diff.path.to_string_lossy().into_owned(),
-                old_text: diff.old_text.clone(),
-                new_text: diff.new_text.clone(),
-            }),
-            SdkToolCallContent::Terminal(_) => None,
-            _ => None,
-        })
-        .collect();
-
-    if items.is_empty() { None } else { Some(items) }
-}
-
-fn map_tool_call_locations(locations: &[SdkToolCallLocation]) -> Option<Vec<AcpToolCallLocationItem>> {
-    (!locations.is_empty()).then(|| {
-        locations
-            .iter()
-            .map(|loc| AcpToolCallLocationItem {
-                path: loc.path.to_string_lossy().into_owned(),
-            })
-            .collect()
-    })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use agent_client_protocol::schema::{
-        ContentBlock, ContentChunk, PermissionOption, PermissionOptionKind as SdkPermissionOptionKind,
-        SessionNotification, SessionUpdate, ToolCall as SdkToolCall, ToolCallStatus as SdkToolCallStatus,
-        ToolCallUpdate as SdkToolCallUpdate, ToolCallUpdateFields, ToolKind as SdkToolKind,
-    };
-    use serde_json::json;
-
-    #[test]
-    fn text_event_roundtrip() {
-        let event = AgentStreamEvent::Text(TextEventData {
-            content: "Hello world".into(),
-        });
-        let json = serde_json::to_value(&event).unwrap();
-        assert_eq!(json["type"], "content");
-        assert_eq!(json["data"]["content"], "Hello world");
-
-        let parsed: AgentStreamEvent = serde_json::from_value(json).unwrap();
-        if let AgentStreamEvent::Text(data) = parsed {
-            assert_eq!(data.content, "Hello world");
-        } else {
-            panic!("Expected Text event");
-        }
-    }
-
-    #[test]
-    fn tips_event_roundtrip() {
-        let event = AgentStreamEvent::Tips(TipsEventData {
-            content: "Something went wrong".into(),
-            tip_type: TipType::Error,
-        });
-        let json = serde_json::to_value(&event).unwrap();
-        assert_eq!(json["type"], "tips");
-        assert_eq!(json["data"]["type"], "error");
-    }
-
-    #[test]
-    fn tool_call_event_roundtrip() {
-        let event = AgentStreamEvent::ToolCall(ToolCallEventData {
-            call_id: "call-1".into(),
-            name: "read_file".into(),
-            args: json!({ "path": "/tmp/a.txt" }),
-            status: ToolCallStatus::Running,
-            input: None,
-            output: None,
-            description: None,
-        });
-        let json = serde_json::to_value(&event).unwrap();
-        assert_eq!(json["type"], "tool_call");
-        assert_eq!(json["data"]["call_id"], "call-1");
-        assert_eq!(json["data"]["status"], "running");
-    }
-
-    #[test]
-    fn tool_call_event_includes_enriched_fields() {
-        let event = AgentStreamEvent::ToolCall(ToolCallEventData {
-            call_id: "call-1".into(),
-            name: "Glob".into(),
-            args: json!({}),
-            status: ToolCallStatus::Completed,
-            input: Some(json!({ "pattern": "**/*.rs" })),
-            output: Some("src/main.rs\nsrc/lib.rs".into()),
-            description: Some("Search for Rust files".into()),
-        });
-        let json = serde_json::to_value(&event).unwrap();
-        assert_eq!(json["type"], "tool_call");
-        assert_eq!(json["data"]["input"]["pattern"], "**/*.rs");
-        assert_eq!(json["data"]["output"], "src/main.rs\nsrc/lib.rs");
-        assert_eq!(json["data"]["description"], "Search for Rust files");
-    }
-
-    #[test]
-    fn tool_call_event_omits_none_fields() {
-        let event = AgentStreamEvent::ToolCall(ToolCallEventData {
-            call_id: "call-1".into(),
-            name: "Glob".into(),
-            args: json!({}),
-            status: ToolCallStatus::Running,
-            input: None,
-            output: None,
-            description: None,
-        });
-        let json = serde_json::to_value(&event).unwrap();
-        assert!(json["data"].get("input").is_none());
-        assert!(json["data"].get("output").is_none());
-        assert!(json["data"].get("description").is_none());
-    }
-
-    #[test]
-    fn finish_event_roundtrip() {
-        let event = AgentStreamEvent::Finish(FinishEventData {
-            session_id: Some("sess-abc".into()),
-        });
-        let json = serde_json::to_value(&event).unwrap();
-        assert_eq!(json["type"], "finish");
-        assert_eq!(json["data"]["session_id"], "sess-abc");
-    }
-
-    #[test]
-    fn error_event_roundtrip() {
-        let event = AgentStreamEvent::Error(ErrorEventData {
-            message: "timeout".into(),
-            code: Some("E001".into()),
-        });
-        let json = serde_json::to_value(&event).unwrap();
-        assert_eq!(json["type"], "error");
-        assert_eq!(json["data"]["message"], "timeout");
-    }
-
-    #[test]
-    fn start_event_default_session_id() {
-        let event = AgentStreamEvent::Start(StartEventData::default());
-        let json = serde_json::to_value(&event).unwrap();
-        assert_eq!(json["type"], "start");
-        assert_eq!(json["data"]["session_id"], serde_json::Value::Null);
-    }
-
-    #[test]
-    fn tool_group_event_roundtrip() {
-        let entries = vec![
-            ToolGroupEntry {
-                call_id: "c1".into(),
-                name: "read".into(),
-                status: ToolCallStatus::Completed,
-                description: Some("Read file".into()),
-            },
-            ToolGroupEntry {
-                call_id: "c2".into(),
-                name: "write".into(),
-                status: ToolCallStatus::Running,
-                description: None,
-            },
-        ];
-        let event = AgentStreamEvent::ToolGroup(entries);
-        let json = serde_json::to_value(&event).unwrap();
-        assert_eq!(json["type"], "tool_group");
-        let data = json["data"].as_array().unwrap();
-        assert_eq!(data.len(), 2);
-        assert_eq!(data[0]["call_id"], "c1");
-    }
-
-    #[test]
-    fn agent_status_event_roundtrip() {
-        let event = AgentStreamEvent::AgentStatus(AgentStatusEventData {
-            backend: "claude".into(),
-            status: "running".into(),
-            agent_name: Some("default".into()),
-            session_id: None,
-        });
-        let json = serde_json::to_value(&event).unwrap();
-        assert_eq!(json["type"], "agent_status");
-        assert_eq!(json["data"]["backend"], "claude");
-    }
-
-    #[test]
-    fn session_tool_call_maps_to_acp_tool_call_event() {
-        let notif = SessionNotification::new(
-            "sess-1",
-            SessionUpdate::ToolCall(
-                SdkToolCall::new("tool-1", "Terminal")
-                    .kind(SdkToolKind::Execute)
-                    .status(SdkToolCallStatus::Pending)
-                    .raw_input(json!({ "command": "echo hi" })),
-            ),
-        );
-
-        let events = session_notification_to_events(&notif);
-        assert_eq!(events.len(), 1);
-        let json = serde_json::to_value(&events[0]).unwrap();
-        assert_eq!(json["type"], "acp_tool_call");
-        assert_eq!(json["data"]["session_id"], "sess-1");
-        assert_eq!(json["data"]["update"]["sessionUpdate"], "tool_call");
-        assert_eq!(json["data"]["update"]["tool_call_id"], "tool-1");
-        assert_eq!(json["data"]["update"]["title"], "Terminal");
-        assert_eq!(json["data"]["update"]["kind"], "execute");
-        assert_eq!(json["data"]["update"]["rawInput"]["command"], "echo hi");
-    }
-
-    #[test]
-    fn session_tool_call_update_omits_missing_fields_for_frontend_merge() {
-        let notif = SessionNotification::new(
-            "sess-1",
-            SessionUpdate::ToolCallUpdate(SdkToolCallUpdate::new(
-                "tool-1",
-                ToolCallUpdateFields::new().status(SdkToolCallStatus::Completed),
-            )),
-        );
-
-        let events = session_notification_to_events(&notif);
-        assert_eq!(events.len(), 1);
-        let json = serde_json::to_value(&events[0]).unwrap();
-        assert_eq!(json["type"], "acp_tool_call");
-        assert_eq!(json["data"]["update"]["sessionUpdate"], "tool_call_update");
-        assert_eq!(json["data"]["update"]["tool_call_id"], "tool-1");
-        assert_eq!(json["data"]["update"]["status"], "completed");
-        assert!(json["data"]["update"].get("title").is_none());
-        assert!(json["data"]["update"].get("rawInput").is_none());
-    }
-
-    #[test]
-    fn permission_request_maps_to_snake_case_event_data() {
-        let request = RequestPermissionRequest::new(
-            "sess-1",
-            SdkToolCallUpdate::new(
-                "tool-1",
-                ToolCallUpdateFields::new()
-                    .title("Write file")
-                    .kind(SdkToolKind::Edit)
-                    .raw_input(json!({ "file_path": "/tmp/a.txt" })),
-            ),
-            vec![
-                PermissionOption::new("allow", "Allow", SdkPermissionOptionKind::AllowOnce),
-                PermissionOption::new("reject", "Reject", SdkPermissionOptionKind::RejectOnce),
-            ],
-        );
-
-        let event = AgentStreamEvent::AcpPermission(permission_request_to_event_data(&request));
-        let json = serde_json::to_value(&event).unwrap();
-
-        assert_eq!(json["type"], "acp_permission");
-        assert_eq!(json["data"]["session_id"], "sess-1");
-        assert_eq!(json["data"]["tool_call"]["tool_call_id"], "tool-1");
-        assert_eq!(json["data"]["tool_call"]["raw_input"]["file_path"], "/tmp/a.txt");
-        assert_eq!(json["data"]["options"][0]["option_id"], "allow");
-        assert_eq!(json["data"]["options"][0]["kind"], "allow_once");
-        assert!(json["data"].get("toolCall").is_none());
-        assert!(json["data"]["options"][0].get("optionId").is_none());
-    }
-
-    #[test]
-    fn thinking_event_roundtrip() {
-        let event = AgentStreamEvent::Thinking(ThinkingEventData {
-            content: "Analyzing...".into(),
-            subject: Some("code review".into()),
-            duration: Some(1500),
-            status: Some("in_progress".into()),
-        });
-        let json = serde_json::to_value(&event).unwrap();
-        assert_eq!(json["type"], "thinking");
-        assert_eq!(json["data"]["duration"], 1500);
-    }
-
-    // ── session_notification_to_chunks projection tests ──────────────────────
-
-    #[test]
-    fn agent_message_chunk_maps_to_text_chunk() {
-        let notif = SessionNotification::new(
-            "sess-1",
-            SessionUpdate::AgentMessageChunk(ContentChunk::new(ContentBlock::from("hello world"))),
-        );
-        let chunks = session_notification_to_chunks(&notif);
-        assert_eq!(chunks.len(), 1);
-        match &chunks[0] {
-            AgentStreamChunk::Text { text } => assert_eq!(text, "hello world"),
-            other => panic!("expected Text, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn agent_thought_chunk_maps_to_thought_chunk() {
-        let notif = SessionNotification::new(
-            "sess-1",
-            SessionUpdate::AgentThoughtChunk(ContentChunk::new(ContentBlock::from("thinking..."))),
-        );
-        let chunks = session_notification_to_chunks(&notif);
-        assert_eq!(chunks.len(), 1);
-        match &chunks[0] {
-            AgentStreamChunk::Thought { content } => assert_eq!(content, "thinking..."),
-            other => panic!("expected Thought, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn tool_call_maps_to_tool_use_chunk() {
-        let notif = SessionNotification::new(
-            "sess-1",
-            SessionUpdate::ToolCall(
-                SdkToolCall::new("tool-1", "read_file")
-                    .kind(SdkToolKind::Read)
-                    .status(SdkToolCallStatus::Pending)
-                    .raw_input(json!({ "path": "/tmp/a.txt" })),
-            ),
-        );
-        let chunks = session_notification_to_chunks(&notif);
-        assert_eq!(chunks.len(), 1);
-        match &chunks[0] {
-            AgentStreamChunk::ToolUse { tool_name, input } => {
-                assert_eq!(tool_name, "read_file");
-                assert_eq!(input["path"], "/tmp/a.txt");
-            }
-            other => panic!("expected ToolUse, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn tool_call_update_without_raw_input_is_skipped() {
-        // Status-only updates should not produce a ToolUse chunk — they'd
-        // duplicate earlier emits without adding signal for the watchdog.
-        let notif = SessionNotification::new(
-            "sess-1",
-            SessionUpdate::ToolCallUpdate(SdkToolCallUpdate::new(
-                "tool-1",
-                ToolCallUpdateFields::new().status(SdkToolCallStatus::Completed),
-            )),
-        );
-        let chunks = session_notification_to_chunks(&notif);
-        assert!(chunks.is_empty());
-    }
-
-    #[test]
-    fn tool_call_update_with_raw_input_emits_tool_use() {
-        let notif = SessionNotification::new(
-            "sess-1",
-            SessionUpdate::ToolCallUpdate(SdkToolCallUpdate::new(
-                "tool-1",
-                ToolCallUpdateFields::new()
-                    .title("search")
-                    .raw_input(json!({ "q": "foo" })),
-            )),
-        );
-        let chunks = session_notification_to_chunks(&notif);
-        assert_eq!(chunks.len(), 1);
-        match &chunks[0] {
-            AgentStreamChunk::ToolUse { tool_name, input } => {
-                assert_eq!(tool_name, "search");
-                assert_eq!(input["q"], "foo");
-            }
-            other => panic!("expected ToolUse, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn plan_update_does_not_produce_chunks() {
-        // Handshake-like updates (plan/mode/model/config/session-info/usage)
-        // should not surface on the subscribe_stream channel.
-        let notif = SessionNotification::new(
-            "sess-1",
-            SessionUpdate::CurrentModeUpdate(agent_client_protocol::schema::CurrentModeUpdate::new("normal")),
-        );
-        let chunks = session_notification_to_chunks(&notif);
-        assert!(chunks.is_empty());
-    }
-}
+pub use crate::protocol::events::*;
